requirements.txt
fastapi
uvicorn
python-dotenv
murf
pydantic
jinja2
assemblyai